v.0.1.0 Major parts ported over
Most classes in the Java version of this project have been ported over. There is still lot's of compatibility issues and the code will not compile in its current state due to not having correct syntax and missing some code. Version number will be aligned with the java version once it has all been ported over.